Add mach bootstrap-android and test-android-startup commands

This adds a `./mach boostrap-android` subcommand that downloads and installs the tools, SDK, NDK,  emulator, and system image for Android. In an environment that can build Servo at all, this should be enough to get all additional dependencies to cross-compile to Android, package an APK, and load it onto a device or an emulator.

At the moment it requires an interactive user to accept the license (and confirm no customization of the emulated virtual device hardware), and then prints environment variables to set for `mach` as well as the command to run to start the emulator (with an already-configured image). A possible next step could be to automate all this, and have `./mach build` run it automatically when needed. (I don’t know if auto-accepting the license is something we should do though.)

This also adds `--emulator` and `--usb` parameters to `./mach install --android` and `./mach run --android`, which tell `adb` what device to pick when both are present. And makes `./mach run --android` print the new process’s PID, for use with e.g. `adb -e logcat --pid 2263`.

Finally, adds the `./mach boostrap-android` subcommand which starts an emulator, installs the APK (it assumes that `mach build` and `mach package` were already executed), runs a single HTML test case, and checks for a message coming from JS through `console.log()` and `adb logcat`.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/servo/servo/21094)
<!-- Reviewable:end -->

+++ python/servo/util.py
@@ -9,6 +9,7 @@
 
 from __future__ import absolute_import, print_function, unicode_literals
 
+import hashlib
 import os
 import os.path
 import platform
@@ -153,9 +154,26 @@ def download_file(desc, src, dst):
     os.rename(tmp_path, dst)
 
 
-def extract(src, dst, movedir=None):
+# https://stackoverflow.com/questions/39296101/python-zipfile-removes-execute-permissions-from-binaries
+# In particular, we want the executable bit for executable files.
+class ZipFileWithUnixPermissions(zipfile.ZipFile):
+    def extract(self, member, path=None, pwd=None):
+        if not isinstance(member, zipfile.ZipInfo):
+            member = self.getinfo(member)
+
+        if path is None:
+            path = os.getcwd()
+
+        extracted = self._extract_member(member, path, pwd)
+        mode = os.stat(extracted).st_mode
+        mode |= (member.external_attr >> 16)
+        os.chmod(extracted, mode)
+        return extracted
+
+
+def extract(src, dst, movedir=None, remove=True):
     assert src.endswith(".zip")
-    zipfile.ZipFile(src).extractall(dst)
+    ZipFileWithUnixPermissions(src).extractall(dst)
 
     if movedir:
         for f in os.listdir(movedir):
@@ -164,4 +182,18 @@ def extract(src, dst, movedir=None):
             os.rename(frm, to)
         os.rmdir(movedir)
 
-    os.remove(src)
+    if remove:
+        os.remove(src)
+
+
+def check_hash(filename, expected, algorithm):
+    hasher = hashlib.new(algorithm)
+    with open(filename, "rb") as f:
+        while True:
+            block = f.read(16 * 1024)
+            if len(block) == 0:
+                break
+            hasher.update(block)
+    if hasher.hexdigest() != expected:
+        print("Incorrect {} hash for {}".format(algorithm, filename))
+        sys.exit(1)
